feat: 메모 불러오기 함수를 UI에 적용
