+++ src/app/src/main/java/com/imeanttobe/app901/ui/component/ImportFromRecipeDialog.kt
@@ -1,5 +1,6 @@
 package com.imeanttobe.app901.ui.component
 
+import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -13,14 +14,17 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.AutoGraph
 import androidx.compose.material.icons.rounded.Clear
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.LoadingIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -33,10 +37,13 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.imeanttobe.app901.R
+import com.imeanttobe.app901.data.type.ConcurrencyState
 
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
 fun ImportFromRecipeDialog(
     urlFieldContent: String,
+    concurrencyState: ConcurrencyState,
     onUrlChange: (String) -> Unit,
     onDismiss: () -> Unit,
     onConfirm: () -> Unit,
@@ -44,6 +51,15 @@ fun ImportFromRecipeDialog(
     val context = LocalContext.current
     val radioOptions = listOf(context.getString(R.string.youtube_shorts), context.getString(R.string.naver_blog))
     val (selectedOption, onOptionSelected) = remember { mutableStateOf(radioOptions[0]) }
+
+    LaunchedEffect(key1 = concurrencyState) {
+        if (concurrencyState is ConcurrencyState.Success) {
+            onDismiss()
+        } else if (concurrencyState is ConcurrencyState.Failure) {
+            onDismiss()
+            Toast.makeText(context, concurrencyState.message, Toast.LENGTH_SHORT).show()
+        }
+    }
 
     AlertDialog(
         icon = {
@@ -59,56 +75,60 @@ fun ImportFromRecipeDialog(
             )
         },
         text = {
-            Column {
-                Text(text = stringResource(R.string.tips_import_from_recipe))
-                Spacer(modifier = Modifier.height(16.dp))
+            if (concurrencyState == ConcurrencyState.Default) {
+                Column {
+                    Text(text = stringResource(R.string.tips_import_from_recipe))
+                    Spacer(modifier = Modifier.height(16.dp))
 
-                radioOptions.forEach { text ->
-                    Row(
-                        Modifier
-                            .clip(RoundedCornerShape(100.dp))
-                            .selectable(
+                    radioOptions.forEach { text ->
+                        Row(
+                            Modifier
+                                .clip(RoundedCornerShape(100.dp))
+                                .selectable(
+                                    selected = (text == selectedOption),
+                                    onClick = { onOptionSelected(text) },
+                                    role = Role.RadioButton,
+                                ).padding(vertical = 8.dp, horizontal = 16.dp),
+                            verticalAlignment = Alignment.CenterVertically,
+                        ) {
+                            RadioButton(
                                 selected = (text == selectedOption),
-                                onClick = { onOptionSelected(text) },
-                                role = Role.RadioButton,
-                            ).padding(vertical = 8.dp, horizontal = 16.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        RadioButton(
-                            selected = (text == selectedOption),
-                            onClick = null, // null recommended for accessibility with screen readers
-                        )
-                        Text(
-                            text = text,
-                            style = MaterialTheme.typography.bodyLarge,
-                            modifier = Modifier.padding(start = 16.dp),
-                        )
-                        if (text != radioOptions.last()) {
-                            Spacer(modifier = Modifier.height(4.dp))
+                                onClick = null, // null recommended for accessibility with screen readers
+                            )
+                            Text(
+                                text = text,
+                                style = MaterialTheme.typography.bodyLarge,
+                                modifier = Modifier.padding(start = 16.dp),
+                            )
+                            if (text != radioOptions.last()) {
+                                Spacer(modifier = Modifier.height(4.dp))
+                            }
                         }
                     }
-                }
 
-                Spacer(modifier = Modifier.height(16.dp))
-                OutlinedTextField(
-                    value = urlFieldContent,
-                    onValueChange = { newValue -> onUrlChange(newValue.replace("\n", "")) },
-                    maxLines = 1,
-                    singleLine = true,
-                    placeholder = { Text(text = stringResource(id = R.string.example_import_from_recipe)) },
-                    keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done, keyboardType = KeyboardType.Uri),
-                    trailingIcon = {
-                        IconButton(
-                            onClick = { onUrlChange("") },
-                        ) {
-                            Icon(
-                                imageVector = Icons.Rounded.Clear,
-                                contentDescription = "Clear",
-                            )
-                        }
-                    },
-                    modifier = Modifier.fillMaxWidth(),
-                )
+                    Spacer(modifier = Modifier.height(16.dp))
+                    OutlinedTextField(
+                        value = urlFieldContent,
+                        onValueChange = { newValue -> onUrlChange(newValue.replace("\n", "")) },
+                        maxLines = 1,
+                        singleLine = true,
+                        placeholder = { Text(text = stringResource(id = R.string.example_import_from_recipe)) },
+                        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done, keyboardType = KeyboardType.Uri),
+                        trailingIcon = {
+                            IconButton(
+                                onClick = { onUrlChange("") },
+                            ) {
+                                Icon(
+                                    imageVector = Icons.Rounded.Clear,
+                                    contentDescription = "Clear",
+                                )
+                            }
+                        },
+                        modifier = Modifier.fillMaxWidth(),
+                    )
+                }
+            } else {
+                LoadingIndicator()
             }
         },
         confirmButton = {
