+++ src/app/src/main/java/com/imeanttobe/app901/ui/home/HomePage.kt
@@ -9,6 +9,7 @@ import androidx.compose.material.icons.rounded.AutoGraph
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
@@ -16,6 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.imeanttobe.app901.R
 import com.imeanttobe.app901.data.enum.HomePageDialogState
+import com.imeanttobe.app901.data.type.ConcurrencyState
 import com.imeanttobe.app901.navigation.BottomNavItem
 import com.imeanttobe.app901.navigation.NavItem
 import com.imeanttobe.app901.ui.component.BottomBar
@@ -26,6 +28,7 @@ import com.imeanttobe.app901.ui.history.HistorySection
 import com.imeanttobe.app901.ui.memo.MemoSection
 import com.imeanttobe.app901.ui.memo.component.MemoFloatingActionButtonMenu
 import com.imeanttobe.app901.ui.profile.ProfileSection
+import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
@@ -48,6 +51,16 @@ fun HomePage(
                 { viewModel.setDialogState(HomePageDialogState.IMPORT_FROM_RECIPE) },
             ),
         )
+
+    LaunchedEffect(key1 = viewModel.importFromUrlConcurrencyState.value) {
+        if (viewModel.importFromUrlConcurrencyState.value is ConcurrencyState.Success) {
+            delay(300)
+            viewModel.resetImportFromUrlConcurrencyState()
+        } else if (viewModel.importFromUrlConcurrencyState.value is ConcurrencyState.Failure) {
+            delay(300)
+            viewModel.resetImportFromUrlConcurrencyState()
+        }
+    }
 
     Scaffold(
         topBar = {},
@@ -98,7 +111,8 @@ fun HomePage(
                     urlFieldContent = viewModel.urlDialogText.value,
                     onUrlChange = { newValue -> viewModel.setUrlDialogText(newValue) },
                     onDismiss = { viewModel.setDialogState(HomePageDialogState.NONE) },
-                    onConfirm = { }, // TODO: Have to attach API here
+                    concurrencyState = viewModel.importFromUrlConcurrencyState.value,
+                    onConfirm = { viewModel.importMemosFromUrl("https://www.naver.com") },
                 )
             HomePageDialogState.NONE -> null
         }
